Desktop/heartpin/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
faiss-cpu>=1.7.4
llama-index>=0.10.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
sentence-transformers>=2.2.0
